@@ -1,51 +1,56 @@
 {
-  "name": "roots/sage",
-  "type": "wordpress-theme",
-  "license": "MIT",
-  "description": "WordPress starter theme with a modern development workflow",
-  "homepage": "https://roots.io/sage/",
-  "authors": [
-    {
-      "name": "Ben Word",
-      "email": "ben@benword.com",
-      "homepage": "https://github.com/retlehs"
+    "name": "roots/sage",
+    "type": "wordpress-theme",
+    "license": "MIT",
+    "description": "WordPress starter theme with a modern development workflow",
+    "homepage": "https://roots.io/sage/",
+    "authors": [
+        {
+            "name": "Ben Word",
+            "email": "ben@benword.com",
+            "homepage": "https://github.com/retlehs"
+        },
+        {
+            "name": "Scott Walkinshaw",
+            "email": "scott.walkinshaw@gmail.com",
+            "homepage": "https://github.com/swalkinshaw"
+        },
+        {
+            "name": "QWp6t",
+            "email": "hi@qwp6t.me",
+            "homepage": "https://github.com/qwp6t"
+        }
+    ],
+    "keywords": [
+        "wordpress"
+    ],
+    "support": {
+        "issues": "https://github.com/roots/sage/issues",
+        "forum": "https://discourse.roots.io/"
     },
-    {
-      "name": "Scott Walkinshaw",
-      "email": "scott.walkinshaw@gmail.com",
-      "homepage": "https://github.com/swalkinshaw"
+    "autoload": {
+        "psr-4": {
+            "App\\": "app/"
+        }
     },
-    {
-      "name": "QWp6t",
-      "email": "hi@qwp6t.me",
-      "homepage": "https://github.com/qwp6t"
+    "require": {
+        "php": ">=7.4",
+        "composer/installers": "~1.0",
+        "illuminate/support": "^5.8.0",
+        "roots/sage-lib": "~9.0.9",
+        "soberwp/controller": "~2.1.0"
+    },
+    "require-dev": {
+        "squizlabs/php_codesniffer": "^2.8.0",
+        "roots/sage-installer": "~1.4",
+        "composer/composer": "^2.1"
+    },
+    "scripts": {
+        "test": [
+            "phpcs"
+        ],
+        "post-create-project-cmd": [
+            "Roots\\Sage\\Installer\\ComposerScript::postCreateProject"
+        ]
     }
-  ],
-  "keywords": ["wordpress"],
-  "support": {
-    "issues": "https://github.com/roots/sage/issues",
-    "forum": "https://discourse.roots.io/"
-  },
-  "autoload": {
-    "psr-4": {
-      "App\\": "app/"
-    }
-  },
-  "require": {
-    "php": ">=7.3",
-    "composer/installers": "~1.0",
-    "illuminate/support": "5.6.*",
-    "roots/sage-lib": "~9.0.9",
-    "soberwp/controller": "~2.1.0"
-  },
-  "require-dev": {
-    "squizlabs/php_codesniffer": "^2.8.0",
-    "roots/sage-installer": "~1.4"
-  },
-  "scripts": {
-    "test": ["phpcs"],
-    "post-create-project-cmd": [
-      "Roots\\Sage\\Installer\\ComposerScript::postCreateProject"
-    ]
-  }
 }
