@@ -6,14 +6,14 @@ use Sober\Controller\Controller;
 
 class App extends Controller
 {
-	
-	/**
-	 * ACF Field Values
-	 *
-	 * @var string[]
-	 */
-	protected $acf = [];
-	
+
+    /**
+     * ACF Field Values
+     *
+     * @var string[]
+     */
+    protected $acf = [];
+
     public function siteName()
     {
         return get_bloginfo('name');
@@ -22,8 +22,9 @@ class App extends Controller
     public static function title()
     {
         if (is_home()) {
-            if ($home = get_option('page_for_posts', true)) {
-                return get_the_title($home);
+            $homepage = get_option('page_for_posts', true);
+            if ($homepage) {
+                return get_the_title($homepage);
             }
             return __('Latest Posts', 'sage');
         }
