@@ -6,19 +6,18 @@ use Sober\Controller\Controller;
 
 class FourZeroFour extends Controller
 {
-	
-	/**
-	 * Dedicated 404 Controller
-	 *
-	 * @var string
-	 */
-	protected $template = '404';
-	
-	/**
-	 * ACF Field Values
-	 *
-	 * @var string[]
-	 */
-	protected $acf = [];
-	
+    
+    /**
+     * Dedicated 404 Controller
+     *
+     * @var string
+     */
+    protected $template = '404';
+    
+    /**
+     * ACF Field Values
+     *
+     * @var string[]
+     */
+    protected $acf = [];
 }
