@@ -6,12 +6,11 @@ use Sober\Controller\Controller;
 
 class FrontPage extends Controller
 {
-	
-	/**
-	 * ACF Field Values
-	 *
-	 * @var string[]
-	 */
-	protected $acf = [];
-	
+    
+    /**
+     * ACF Field Values
+     *
+     * @var string[]
+     */
+    protected $acf = [];
 }
