@@ -3,16 +3,14 @@
 namespace App\Controllers;
 
 use Sober\Controller\Controller;
-use function App\grab_excerpt;
 
 class Single extends Controller
 {
-	
-	/**
-	 * ACF Field Values
-	 *
-	 * @var string[]
-	 */
-	protected $acf = [];
-	
+
+    /**
+     * ACF Field Values
+     *
+     * @var string[]
+     */
+    protected $acf = [];
 }
