@@ -1,10 +1,12 @@
 <?php
+
 declare(strict_types = 1);
 
 namespace App\Helper;
 
 class ImageHelper
 {
+
     /**
      * Returns a picture tag serving webp images in a source tag for an image
      *
@@ -19,7 +21,7 @@ class ImageHelper
      *
      * @return string          Returns a formatted picture HTML element with source and img tags
      */
-    public static function imgSrcSet($src_id, $img_attrs = [], $data_atts = [], $aria_atts = [])
+    public static function imgSrcSet(int $src_id, array $img_attrs = [], array $data_atts = [], array $aria_atts = [])
     {
         $get_img_data = self::imgSrcSetArr($src_id, $img_attrs);
 
@@ -27,17 +29,17 @@ class ImageHelper
         $da_return = [];
         $aa_return = [];
 
-        $prop        = $get_img_data['prop'];
+        $properties  = $get_img_data['prop'];
         $image_class = $img_attrs['class'] ?? '';
         $image_id    = $img_attrs['id'] ?? '';
 
         // Image Information
-        $src       = $get_img_data['url'] ?: '';
+        $source    = $get_img_data['url'] ?: '';
         $src_set   = $get_img_data['src_set'] ?: '';
         $src_sizes = $get_img_data['src_set_sizes'] ?? '';
         $get_webp  = $get_img_data['webp'] ?? '';
 
-        $role = $img_attrs['role'] ?? ($get_img_data['role'] ?? 'img');
+        $role_atts = $img_attrs['role'] ?? ($get_img_data['role'] ?? 'img');
 
         if ($data_atts) {
             $da_return = collect($data_atts)
@@ -77,29 +79,29 @@ class ImageHelper
         if (strpos($get_img_data['img_type'], 'svg')) {
             return sprintf(
                 '<picture><img src="%1$s" role="%4$s" %2$s property="v:%3$s" %5$s %6$s content="%1$s" %7$s %8$s/></picture>',
-                esc_url($src),
+                esc_url($source),
                 $get_img_data['alt'] ? "alt=\"{$get_img_data['alt']}\"" : '',
-                $prop,
-                $role,
-                $image_class ? "class=\"{$image_class}\"" : '',
-                $image_id ? "id=\"{$image_id}\"" : '',
-                implode(' ', $da_return),
-                implode(' ', $aa_return),
-            );
-        } else {
-            return sprintf(
-                '<picture>%2$s<img role="%5$s" %3$s property="v:%4$s" %6$s %7$s content="%1$s" %8$s %9$s/></picture>',
-                esc_url($src),
-                ($webp_source ?? '') . $og_source,
-                $get_img_data['alt'] ? "alt=\"{$get_img_data['alt']}\"" : '',
-                $prop,
-                $role,
+                $properties,
+                $role_atts,
                 $image_class ? "class=\"{$image_class}\"" : '',
                 $image_id ? "id=\"{$image_id}\"" : '',
                 implode(' ', $da_return),
                 implode(' ', $aa_return),
             );
         }
+
+        return sprintf(
+            '<picture>%2$s<img role="%5$s" %3$s property="v:%4$s" %6$s %7$s content="%1$s" %8$s %9$s/></picture>',
+            esc_url($source),
+            ($webp_source ?? '') . $og_source,
+            $get_img_data['alt'] ? "alt=\"{$get_img_data['alt']}\"" : '',
+            $properties,
+            $role_atts,
+            $image_class ? "class=\"{$image_class}\"" : '',
+            $image_id ? "id=\"{$image_id}\"" : '',
+            implode(' ', $da_return),
+            implode(' ', $aa_return),
+        );
     }
 
     /**
@@ -111,7 +113,7 @@ class ImageHelper
      *
      * @return string
      */
-    public static function imgSrcSetCaption($src_id, $img_attrs = [], $data_atts = [], $aria_atts = [])
+    public static function imgSrcSetCaption(int $src_id, array $img_attrs = [], array $data_atts = [], array $aria_atts = [])
     {
 
         $get_img_data = self::imgSrcSetArr($src_id, $img_attrs);
@@ -121,18 +123,18 @@ class ImageHelper
         $aa_return = [];
 
         // $img_attrs
-        $prop        = $get_img_data['prop'];
+        $properties        = $get_img_data['prop'];
         $image_class = $img_attrs['class'] ?? '';
         $image_id    = $img_attrs['id'] ?? '';
         $fig_class   = $img_attrs['figure_class'] ?? '';
 
         // Image Information
-        $src       = $get_img_data['url'] ?: '';
+        $source       = $get_img_data['url'] ?: '';
         $src_set   = $get_img_data['src_set'] ?: '';
         $src_sizes = $get_img_data['src_set_sizes'] ?? '';
         $src_alt   = $img_attrs['alt'] ?? '';
         $get_webp  = $get_img_data['webp'] ?? '';
-        $role      = $img_attrs['role'] ?? ($get_img_data['role'] ?? 'img');
+        $role_atts      = $img_attrs['role'] ?? ($get_img_data['role'] ?? 'img');
 
         $caption = $img_attrs['caption'] ?? false;
 
@@ -182,30 +184,15 @@ class ImageHelper
             '<source srcset="%s" sizes="%s" type="%s">',
             $src_set,
             $src_sizes,
-            str_replace('.', '', substr($src, strrpos($src, '.')))
+            str_replace('.', '', substr($source, strrpos($source, '.')))
         ) : '';
         if (strpos($get_img_data['img_type'], 'svg')) {
             sprintf(
                 '<figure %5$s><picture><img src="%1$s" role="%4$s" %2$s %6$s %7$s property="v:%3$s" content="%1$s" %8$s %9$s /></picture>%10$s</figure>',
-                esc_url($src),
+                esc_url($source),
                 $src_attributes,
-                $prop,
-                $role,
-                $fig_class ? 'class="wp-caption ' . $fig_class . '"' : '',
-                $image_class ? 'class="' . $image_class . '"' : '',
-                $image_id ? "id=\"{$image_id}\"" : '',
-                implode(' ', $da_return),
-                implode(' ', $aa_return),
-                $caption
-            );
-        } else {
-            return sprintf(
-                '<figure %6$s><picture>%2$s<img role="%5$s" %3$s  %7$s %8$s property="v:%4$s" content="%1$s" %9$s %10$s /></picture>%11$s</figure>',
-                esc_url($src),
-                ($webp_source ?? '') . $og_source,
-                $src_attributes,
-                $prop,
-                $role,
+                $properties,
+                $role_atts,
                 $fig_class ? 'class="wp-caption ' . $fig_class . '"' : '',
                 $image_class ? 'class="' . $image_class . '"' : '',
                 $image_id ? "id=\"{$image_id}\"" : '',
@@ -214,6 +201,20 @@ class ImageHelper
                 $caption
             );
         }
+        return sprintf(
+            '<figure %6$s><picture>%2$s<img role="%5$s" %3$s  %7$s %8$s property="v:%4$s" content="%1$s" %9$s %10$s /></picture>%11$s</figure>',
+            esc_url($source),
+            ($webp_source ?? '') . $og_source,
+            $src_attributes,
+            $properties,
+            $role_atts,
+            $fig_class ? 'class="wp-caption ' . $fig_class . '"' : '',
+            $image_class ? 'class="' . $image_class . '"' : '',
+            $image_id ? "id=\"{$image_id}\"" : '',
+            implode(' ', $da_return),
+            implode(' ', $aa_return),
+            $caption
+        );
     }
 
     /**
@@ -225,7 +226,7 @@ class ImageHelper
      *
      * @return array
      */
-    public static function imgSrcSetArr($src_id, $img_attrs = [])
+    public static function imgSrcSetArr(int $src_id, array $img_attrs = [])
     {
 
         // $img_attrs
@@ -268,13 +269,14 @@ class ImageHelper
      *
      * @return array|string
      */
-    public static function generateImgData($post_thmb_id)
+    public static function generateImgData(int $post_thmb_id)
     {
         if (!$post_thmb_id) {
             return '';
         }
 
-        $img_data['alt']    = get_post_meta($post_thmb_id, '_wp_attachment_image_alt', true) ? html_entity_decode(get_post_meta($post_thmb_id, '_wp_attachment_image_alt', true)) : html_entity_decode(get_the_title($post_thmb_id));
+        $post_meta_alt = get_post_meta($post_thmb_id, '_wp_attachment_image_alt', true);
+        $img_data['alt']    = $post_meta_alt ? html_entity_decode($post_meta_alt) : html_entity_decode(get_the_title($post_thmb_id));
         $img_data['src']    = wp_get_attachment_image_url($post_thmb_id, 'large', false);
         $img_data['srcset'] = wp_get_attachment_image_srcset($post_thmb_id, 'large');
         $img_data['sizes']  = wp_get_attachment_image_sizes($post_thmb_id, 'large') ?: '';
@@ -283,7 +285,7 @@ class ImageHelper
         $img_data['type'] = 'image/' . str_replace('.', '', substr($source, strrpos($source, '.')));
 
         // Handles the webp portion of the image
-        self::getWebp(wp_get_attachment_image_srcset($post_thmb_id, 'large')) ? $img_data['webpSrcset'] = self::getWebp(wp_get_attachment_image_srcset($post_thmb_id, 'large')) : null;
+        self::getWebp($img_data['srcset']) ? $img_data['webpSrcset'] = self::getWebp($img_data['srcset']) : null;
 
         return $img_data;
     }
@@ -296,14 +298,14 @@ class ImageHelper
      *
      * @return string
      */
-    protected static function getWebp($src_set)
+    protected static function getWebp(string $src_set)
     {
 
         // Creation of webp support
         $webp_srcset = [];
         $sset_items  = explode(', ', $src_set);
         foreach ($sset_items as $srcset) {
-            $img_file  = explode(' ', $srcset) ? explode(' ', $srcset) : $srcset;
+            $img_file = explode(' ', $srcset) ? explode(' ', $srcset) : $srcset;
             if (!empty($img_file) && $img_file[0]) {
                 $item_size = !empty($img_file) && isset($img_file[1]) ? $img_file[1] : '';
                 $extension = substr($img_file[0], strrpos($img_file[0], '.'));
