@@ -4,7 +4,15 @@ declare(strict_types = 1);
 
 namespace App\Routes;
 
+# Sage Classes
 use App\Routes\Traits\RestRouteParams;
+
+# Wordpress
+use WP_Error;
+use WP_HTTP_Response;
+use WP_REST_Request;
+use WP_REST_Response;
+use WP_REST_Server;
 
 /**
  * Class Name: SageNavRestAPI
@@ -37,24 +45,26 @@ class SageNavRestAPI
          * Grabs the navigation items
          */
         register_rest_route($this->route, '/get-nav', [
-            'methods'  => \WP_REST_Server::READABLE,
+            'methods'  => WP_REST_Server::READABLE,
             'callback' => [$this, 'navWOChildren'],
+            'permission_callback' => '__return_true',
         ]);
 
         register_rest_route($this->route, '/get-nav-with-children', [
-            'methods'  => \WP_REST_Server::READABLE,
+            'methods'  => WP_REST_Server::READABLE,
             'callback' => [$this, 'navWithChildren'],
+            'permission_callback' => '__return_true',
         ]);
     }
 
     /**
      * Returns an array of navigation items without their children
      *
-     * @param \WP_REST_Request $request
+     * @param WP_REST_Request $request
      *
-     * @return mixed|\WP_Error|\WP_HTTP_Response|\WP_REST_Response
+     * @return mixed|WP_Error|WP_HTTP_Response|WP_REST_Response
      */
-    public function navWOChildren(\WP_REST_Request $request)
+    public function navWOChildren(WP_REST_Request $request)
     {
         $return_menu = [];
 
@@ -67,9 +77,9 @@ class SageNavRestAPI
 
         if (!is_wp_error($get_menu) && !empty($get_menu)) {
             $return_menu = collect($get_menu)
-                ->map(function ($item) {
-                    if ((int) $item->menu_item_parent === 0) {
-                        return self::parseNavItem($item);
+                ->map(function ($the_item) {
+                    if ((int) $the_item->menu_item_parent === 0) {
+                        return self::parseNavItem($the_item);
                     }
                     return '';
                 })
@@ -83,11 +93,11 @@ class SageNavRestAPI
     /**
      * Returns an array of navigation items with children
      *
-     * @param \WP_REST_Request $request
+     * @param WP_REST_Request $request
      *
-     * @return mixed|\WP_Error|\WP_HTTP_Response|\WP_REST_Response
+     * @return mixed|WP_Error|WP_HTTP_Response|WP_REST_Response
      */
-    public function navWithChildren(\WP_REST_Request $request)
+    public function navWithChildren(WP_REST_Request $request)
     {
         $return_menu = [];
 
@@ -107,9 +117,9 @@ class SageNavRestAPI
 
             $return_menu = collect($parse_items)
                 ->first()
-                ->map(function ($item) use ($parse_items) {
-                    $item['children'] = $parse_items[$item['id']] ?? [];
-                    return $item;
+                ->map(function ($the_item) use ($parse_items) {
+                    $the_item['children'] = $parse_items[$the_item['id']] ?? [];
+                    return $the_item;
                 })
                 ->toArray();
         }
@@ -125,7 +135,7 @@ class SageNavRestAPI
      *
      * @return array
      */
-    private function parseNavItem($item, $opts = [])
+    private function parseNavItem(object $item, array $opts = [])
     {
         $default_classes = [
             'menu-item',
