@@ -1,5 +1,7 @@
 <?php
 declare(strict_types = 1);
 
+use \App\Routes\BlogPosts;
+
 // Blog Posts REST API
-// new \App\Routes\BlogPosts();
+// new BlogPosts();
