@@ -2,10 +2,12 @@
 
 namespace App;
 
+use WP_Customize_Manager;
+
 /**
  * Theme customizer
  */
-add_action('customize_register', function (\WP_Customize_Manager $wp_customize) {
+add_action('customize_register', function (WP_Customize_Manager $wp_customize) {
     // Add postMessage support
     $wp_customize->get_setting('blogname')->transport = 'postMessage';
     $wp_customize->selective_refresh->add_partial('blogname', [
