@@ -2,10 +2,9 @@
 	<div class="container">
 		{{-- Uncomment for Vue Navigation --}}
 		{{-- Vue Based Navigation --}}
-		{{-- <div id="vue-navigation-container"></div>--}}
+		{{-- <div id="vue-navigation-container"></div> --}}
 		
 		{{-- WordPress default Navigation --}}
-		
 		<a class="brand" href="{{ home_url('/') }}">{{ get_bloginfo('name', 'display') }}</a>
 		<nav class="nav-primary">
 			@if (has_nav_menu('primary_navigation'))
